Try again when IMAP server is unreachable.
From times to times, the IMAP server can be unreachable because of some
network issues.  When this happened, the program caught the exception
and tried to send an email notification.  As the network issue is still
not fixed yet, the same problem occurs again during the mail
notification making the program ends silently this time.

To prevent this, we catch the gaierror and we do not send a mail notif.
This way, the program continues running and will try again later.