Fixed multi-tab client issue.
When a user was using more than 1 client on different tabs of the same browser, the clients were stealing each other messages because the session-id is the same and they could not distinguish to which tab the responses were addressed.  I have added a client_tab_id that I concatenate with the session-id on the server-side to diffentiate the clients.