Changed protocol to support long messages.
Now that the web interface supports file upload, requests can be a lot
longer than before.  Because of that, I changed the communication
protocol between the clients and the server.  Each request from a client
must be terminated by the EOM delimiter (that is "==== EOM ====").  This
way, the server can use a loop to read very long message until the EOM
is met.