Prevented the worker from crashing.
When a worker tried to connect to a translation server that was down,
the worker would stop working and it was impossible to reuse it even
after restarting the translation server.  I catch the exception, log an
error message, and ignore the request.  Doing so also prevents the
worker from dying so that ulterior translations can be processed once
the translation server is restarted.