Encoded splitter_cmd to utf-8.
This prevents a strange bug.  When the multiserver is started
automatically via a boot script, if we try to translate a sentence, it
will crash when running the splitter_cmd.  We must start the multiserver
manually so that i can work normally.  If I encode the splitter_cmd, it
works all the time.  I don't know understand why though...