Fixed recursive mail issue.
At one moment, a delivery error occurred when sending mail to a
notifier.  Because of that, the Google mail handler notified everyone
with a mail that was also sent to the gmail account.  That warning mail
was interpreted as an translation request from the mail-handler.  As the
message could not be digest properly, an error was sent to all the
listeners.  But the same delivery error occurred, provoking a recursive
issue.

To prevent the creation of a gigantic mass of mails, messages with "daemon" in the From: and To: field are kind of
ignored by the system.